# specs/001-graphrag-chatbot-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: GraphRAG Chatbot Platform API
  description: PDF 기반 GraphRAG 챗봇 플랫폼 API
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: 관리자 로그인
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      summary: 현재 사용자 정보 조회
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 사용자 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============ Chatbots ============
  /chatbots:
    get:
      tags: [Chatbots]
      summary: 챗봇 목록 조회
      operationId: listChatbots
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ChatbotStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 챗봇 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatbotListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Chatbots]
      summary: 챗봇 생성
      operationId: createChatbot
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatbotRequest'
      responses:
        '201':
          description: 챗봇 생성 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chatbot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /chatbots/{chatbotId}:
    get:
      tags: [Chatbots]
      summary: 챗봇 상세 조회
      operationId: getChatbot
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/chatbotId'
      responses:
        '200':
          description: 챗봇 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatbotDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Chatbots]
      summary: 챗봇 수정
      operationId: updateChatbot
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/chatbotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChatbotRequest'
      responses:
        '200':
          description: 수정 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chatbot'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Chatbots]
      summary: 챗봇 삭제
      operationId: deleteChatbot
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/chatbotId'
      responses:
        '204':
          description: 삭제 완료
        '404':
          $ref: '#/components/responses/NotFound'
  /chatbots/{chatbotId}/status:
    patch:
      tags: [Chatbots]
      summary: 챗봇 상태 변경 (활성화/비활성화)
      operationId: updateChatbotStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/chatbotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [active, inactive]
      responses:
        '200':
          description: 상태 변경 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chatbot'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============ Documents ============
...
  /chat/{accessUrl}:
    get:
      tags: [Chat]
      summary: 챗봇 정보 조회 (공개)
      operationId: getChatbotPublic
      parameters:
        - name: accessUrl
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 챗봇 공개 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatbotPublic'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    chatbotId:
      name: chatbotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
          description: 토큰 만료 시간 (초)
    AdminUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        created_at:
          type: string
          format: date-time

    # ---- Chatbot ----
    ChatbotStatus:
      type: string
      enum: [active, inactive, processing]
    Persona:
      type: object
      properties:
        tone:
          type: string
          enum: [professional, friendly, formal, casual]
          default: professional
        language:
          type: string
          enum: [ko, en]
          default: ko
        greeting:
          type: string
          maxLength: 500
        fallback_message:
          type: string
          maxLength: 500
    CreateChatbotRequest:
      type: object
      required: [name, access_url]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        access_url:
          type: string
          pattern: '^[a-z0-9-]{3,50}$'
        persona:
          $ref: '#/components/schemas/Persona'
    UpdateChatbotRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        persona:
          $ref: '#/components/schemas/Persona'
    Chatbot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        persona:
          $ref: '#/components/schemas/Persona'
        status:
          $ref: '#/components/schemas/ChatbotStatus'
        access_url:
          type: string
        active_version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ChatbotDetail:
      allOf:
        - $ref: '#/components/schemas/Chatbot'
        - type: object
          properties:
            document_count:
              type: integer
            total_sessions:
              type: integer
            total_messages:
              type: integer
    ChatbotPublic:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        persona:
          type: object
          properties:
            greeting:
              type: string
    ChatbotListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Chatbot'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ---- Document ----
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object